+++ server/fitnessScore.h
@@ -13,7 +13,8 @@ void stepFitnessScore();
 // 0 still pending
 // -1 DENIED
 // 1 score ready (and returned in outScore)
-int getFitnessScore( char *inEmail, int *outScore );
+// outScore NOT modified if result not ready
+int getFitnessScore( char *inEmail, float *outScore );
 
 // all string params copied internally
 void logFitnessDeath( char *inEmail, char *inName, int inDisplayID,
